feat: update sync for meta discussion
I added an if statement to determine if the repository is
kratos, hydra, oathkeeper, keto.
If it is the script proceeds as normal.
If not then the repository name gets substituted with "meta".

This should point all synced Ory repositories to meta/discussions.

I tested it in /web and /kratos locally and worked fine.

This is just a first draft, I expect to make some more changes :)

(Ps: also changed sync.yml to include ory/works in the synch action again)